% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.TL.R
\name{multiplot.TL}
\alias{multiplot.TL}
\title{multiplot.TL}
\usage{
multiplot.TL(
  file,
  nomFile = "",
  ech = 1,
  Doseb0 = 90,
  Dosea0 = 0,
  supra = FALSE,
  norm = FALSE,
  plateau = seq(200, 500),
  Temp = seq(26, 599),
  NumInv = "",
  b.check = rep(1, 9),
  a.check = rep(1, 4),
  sup.check = rep(1, 9),
  brate = 0.1,
  arate = 0.1
)
}
\arguments{
\item{file}{\linkS4class{Risoe.BINfileData} (\strong{required}) the BIN/BINX file(s)}

\item{nomFile}{\link{character} (\strong{with default}) name of the BIN/BINX file}

\item{ech}{\link{numeric} (\strong{with default}) the sample number}

\item{Doseb0}{\link{numeric} (\strong{with default})  the reference of beta irradiation (in seconds)}

\item{Dosea0}{\link{numeric} (\strong{with default})  the reference of alpha irradiation (in seconds)}

\item{supra}{\link{logical} (\strong{with default}) TRUE if supra measurements are included. Based on the number of data files, the value is corrected}

\item{norm}{\link{logical} (\strong{with default}) TRUE if the measurements are normalized.}

\item{plateau}{\link{numeric},\link{list} (\strong{with default}) the plateau range for normalization}

\item{Temp}{\link{numeric},\link{list} (\strong{with default}) the temperature range}

\item{NumInv}{\link{character} (\strong{with default}) the inventory number}

\item{b.check}{\link{numeric},\link{list} (\strong{with default}) list of beta curves to plot}

\item{a.check}{\link{numeric},\link{list} (\strong{with default}) list of alpha curves to plot}

\item{sup.check}{\link{numeric},\link{list} (\strong{with default}) list of supralinarity curves to plot}

\item{brate}{\link{numeric} (\strong{with default}) dose rate of the beta source (default value 0.1 Gy/s)}

\item{arate}{\link{numeric} (\strong{with default}) dose rate of the alpha source (default value 0.1 micro-2)}
}
\value{
plot figures
}
\description{
plot for TL measurements
}
\examples{

\dontrun{
library(OSLpack)
file<-OSLpack::ReadFile(n=1)

File<-file$FILE
NFile<-file$NFILE
}

}
