% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiTL}
\alias{multiTL}
\title{multiTL}
\format{
A list with 9 variables: \code{Dose}, \code{df.T},
\code{df.y}, \code{mu_b}, \code{mu_0}, \code{var_b}, \code{var_0},
\code{var_y} and \code{iter}
}
\usage{
multiTL
}
\description{
TL example from RLumModel
a dataset containing multiples TL calculated with RLumModel
call function "model_LuminescenceSignals", model = "Bailey2001"
the irradiation dose is varied and then compared, and parameters for #’ the various Slice functions
}
\keyword{datasets}
