% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTL_fun.R
\name{multiTL_fun}
\alias{multiTL_fun}
\title{multiTL_fun}
\usage{
multiTL_fun(
  irradiation_dose = c(20, 20, 20, 40, 40, 40, 60, 60, 60),
  model = "Bailey2001",
  distri_scatter = "runif(1,s,t)",
  par_scatter = list(s = 0.8, t = 1.2),
  distri_noise = "runif(n,a,b)",
  par_noise = list(a = 0.9, b = 1.1)
)
}
\arguments{
\item{irradiation_dose}{\link[=list]{numeric} (\strong{with default}) irradiation dose}

\item{model}{\link{character} (\strong{with default}) model as RLumModel}

\item{distri_scatter}{\link{character} (\strong{with default}) a distribution function to scatter the glow curves.}

\item{par_scatter}{\link[=list]{numeric} (\strong{with default}) scatter parameters associated to scatter distribution.}

\item{distri_noise}{\link{character} (\strong{with default}) a distribution function to noise the glow curve.}

\item{par_noise}{\link[=list]{numeric} (\strong{with default}) noise parameter associated to the noise distribution.}
}
\value{
a dataset containing multiples TL calculated with RLumModel
}
\description{
multiTL example from RLumModel
}
\examples{

\dontrun{
multiTL_fun()
}

}
