% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plateau4.R
\name{Plateau4}
\alias{Plateau4}
\title{Plateau4}
\usage{
Plateau4(
  Dose,
  df.T,
  df.y,
  Ti = 200,
  Tf = 500,
  k = 1,
  yn = 0,
  n.iter,
  n.burnin = n.iter/2,
  n.thin = max(1, floor(n.iter - n.burnin)/10)
)
}
\arguments{
\item{Dose}{\link{numeric} (\strong{required}) set of the irradiation doses}

\item{df.T}{\link{matrix} (\strong{required}) Luminescence data: the temperatures}

\item{df.y}{\link{matrix} (\strong{required}) Luminescence data; the luminescence signal}

\item{Ti}{\link{numeric} (\strong{with default}) temperature minimum for the plateau}

\item{Tf}{\link{numeric} (\strong{with default}) temperature maximum for the plateau}

\item{k}{\link{numeric} (\strong{required}) the number of parameters in the group
involving sigma}

\item{yn}{\link{numeric} (\strong{with default}) intensity value of the 'true' (natural) Dose (in extrapolation, yn = 0)}

\item{n.iter}{\link{numeric} (\strong{required}) the number of iteration}

\item{n.burnin}{\link{numeric} (\strong{with default})    number of iterations to discard at the beginning (burn in).
Default is n.iter/2, that is, discarding the first half of the simulations.}

\item{n.thin}{\link{numeric} (\strong{with default}) thinning rate. Must be a positive integer. Set n.thin > 1 to save memory and computation time if n.iter is large.
Default is max(1, floor((n.iter-n.burnin) / 10)) which will only thin if there are at least 20 simulations.}
}
\value{
an (r × 5)-matrix (mcmc class),
\tabular{lll}{
\strong{column} \tab \strong{Type} \tab \strong{Description}\cr
\code{alpha} \tab  \code{numeric} \tab intercept\cr
\code{beta} \tab \code{numeric} \tab slope \cr
\code{sigma2} \tab  \code{numeric} \tab variance\cr
\code{T} \tab \code{numeric} \tab Temperature \cr
\code{De} \tab \code{numeric} \tab 'True' Dose value \cr
}
}
\description{
MC Analysis TL (Plateau +gibbs) following gibbs4.R Hickey (2006)
Bayesian approach without expert judgement
additional scalar variable T (Plateau sampler)
}
\examples{
##load data
if(dev.cur()!=1) dev.off()
data(multiTL, envir = environment())
Dose<-multiTL$Dose
df.T<-multiTL$df.T
df.y<-multiTL$df.y
n.iter<-multiTL$n.iter
test<-Plateau4(Dose,df.T,df.y,k=1,n.iter=n.iter)


}
\references{
Gibbs sampler: Hickey, G. L. 2006. « The Linear Calibration Problem: A Bayesian Analysis ». PhD Thesis, PhD dissertation, University of Durham. 1–148. http://www.dur.ac.uk/g.l.hickey/dissertation.pdf.

chapter 6.3.6 and Appendix G.7
}
