% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice1.R
\name{Slice1}
\alias{Slice1}
\title{Slice1}
\usage{
Slice1(Dose, df.T, df.y, mu_b, mu_0, var_b, var_0, var_y, n.iter)
}
\arguments{
\item{Dose}{\link{numeric} (\strong{required}) set of the irradiation doses}

\item{df.T}{\link{matrix} (\strong{required}) Luminescence data: the temperatures}

\item{df.y}{\link{matrix} (\strong{required}) Luminescence data; the luminescence signal}

\item{mu_b}{\link{numeric} (\strong{required}) Prior of the slope's average}

\item{mu_0}{\link{numeric} (\strong{required}) Prior of the true dose's average}

\item{var_b}{\link{numeric} (\strong{required}) Prior of the slope's variance}

\item{var_0}{\link{numeric} (\strong{required}) Prior of the true dose's variance}

\item{var_y}{\link{numeric} (\strong{required}) Prior of the intensity's variance}

\item{n.iter}{\link{numeric} (\strong{required}) the number of iteration}
}
\value{
an (n.iter x 3)-matrix
De 'True' dose value
b slope
T Temperature
}
\description{
MC Analysis TL (slice +gibbs) following gibbs1.R Hickey (2006)
extrapolation y0=0 (gibbs1)
additional scalar variable T (Slice sampler)
}
\examples{
##load data
if(dev.cur()!=1) dev.off()
data(multiTL, envir = environment())
attach(multiTL)
Slice1(Dose,df.T,df.y,mu_b,mu_0,var_b,var_0,var_y,n.iter)
detach(multiTL)


}
\references{
Gibbs sampler: Hickey, G. L. 2006. « The Linear Calibration Problem: A Bayesian Analysis ». PhD Thesis, PhD dissertation, University of Durham. 1–148. http://www.dur.ac.uk/g.l.hickey/dissertation.pdf.

chapter 3 and Appendix G.1

Slice sampler: Neal, R. 2003. « Slice Sampling ». Annals of Statistics 31 (3): 705‑67.
}
