% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice3.R
\name{Slice3}
\alias{Slice3}
\title{Slice3}
\usage{
Slice3(
  Dose,
  df.T,
  df.y,
  mu_0,
  var_0,
  mu_alpha,
  var_alpha,
  mu_beta,
  var_beta,
  y0 = 0,
  a,
  b,
  n.iter
)
}
\arguments{
\item{Dose}{\link{numeric} (\strong{required}) set of the irradiation doses}

\item{df.T}{\link{matrix} (\strong{required}) Luminescence data: the temperatures}

\item{df.y}{\link{matrix} (\strong{required}) Luminescence data; the luminescence signal}

\item{mu_0}{\link{numeric} (\strong{required}) Prior of the true dose's average}

\item{var_0}{\link{numeric} (\strong{required}) Prior of the true dose's variance}

\item{mu_alpha}{\link{numeric} (\strong{required}) Prior of the intercept's average}

\item{var_alpha}{\link{numeric} (\strong{required}) Prior of the intercept's variance}

\item{mu_beta}{\link{numeric} (\strong{required}) Prior of the slope's average}

\item{var_beta}{\link{numeric} (\strong{required}) Prior of the slope's variance}

\item{y0}{\link{numeric} (\strong{with default}) Prior of the luminescence for the true dose (Default y0=0, extrapolation) \strong{need ???}}

\item{a}{\link{numeric} (\strong{required}) the lower end points of the experimental calibration range}

\item{b}{\link{numeric} (\strong{required}) the upper end points of the experimental calibration range}

\item{n.iter}{\link{numeric} (\strong{required}) the number of iteration}
}
\value{
an (r × 5)-matrix,
x0 'True' dose value
sigma2 variance
alpha intercept
beta slope
T Temperature
}
\description{
MC Analysis TL (slice +gibbs) following gibbs3.R Hickey (2006)
}
\details{
additional scalar variable T (Slice sampler)
}
\references{
Gibbs sampler: Hickey, G. L. 2006. « The Linear Calibration Problem: A Bayesian Analysis ». PhD Thesis, PhD dissertation, University of Durham. 1–148. http://www.dur.ac.uk/g.l.hickey/dissertation.pdf.

chapter 5 and Appendix G.6

Slice sampler: Neal, R. 2003. « Slice Sampling ». Annals of Statistics 31 (3): 705‑67.
}
