% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice4.R
\name{Slice4}
\alias{Slice4}
\title{Slice4}
\usage{
Slice4(Dose, df.T, df.y, k = 1, n.iter)
}
\arguments{
\item{Dose}{\link{numeric} (\strong{required}) set of the irradiation doses}

\item{df.T}{\link{matrix} (\strong{required}) Luminescence data: the temperatures}

\item{df.y}{\link{matrix} (\strong{required}) Luminescence data; the luminescence signal}

\item{k}{\link{numeric} (\strong{required}) the number of parameters in the group
involving sigma}

\item{n.iter}{\link{numeric} (\strong{required}) the number of iteration}
}
\value{
an (r × 5)-matrix,
De 'True' dose value
sigma2 variance
alpha intercept
beta slope
T Temperature
}
\description{
MC Analysis TL (slice +gibbs) following gibbs4.R Hickey (2006)
Bayesian approach without expert judgement
additional scalar variable T (Slice sampler)
}
\examples{
##load data
if(dev.cur()!=1) dev.off()
data(multiTL, envir = environment())
Dose<-multiTL$Dose
df.T<-multiTL$df.T
df.y<-multiTL$df.y
n.iter<-multiTL$n.iter
test<-Slice4(Dose,df.T,df.y,k=1,n.iter)


}
\references{
Gibbs sampler: Hickey, G. L. 2006. « The Linear Calibration Problem: A Bayesian Analysis ». PhD Thesis, PhD dissertation, University of Durham. 1–148. http://www.dur.ac.uk/g.l.hickey/dissertation.pdf.

chapter 6.3.6 and Appendix G.7

Slice sampler: Neal, R. 2003. « Slice Sampling ». Annals of Statistics 31 (3): 705‑67.
}
