% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step.Out.R
\name{Step.Out}
\alias{Step.Out}
\title{Step.Out}
\usage{
Step.Out(foo, x0, y, w, m, Rmx = 500)
}
\arguments{
\item{foo}{\link{function} (\strong{required}) function proportional to the density}

\item{x0}{\link{numeric} (\strong{required}) the current point}

\item{y}{\link{numeric} (\strong{required}) the vertical level defining the slice}

\item{w}{\link{numeric} (\strong{required}) estimate of the typical size of a slice}

\item{m}{\link{numeric} (\strong{required}) integer limiting the size of a slice to mw}

\item{Rmx}{\link{numeric} (\strong{required}) bound value to limit edge effects (added)}
}
\value{

}
\description{
Slice sampler: stepping out procedure
Neal(2003) fig.3
}
