% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLpack-package.R
\docType{data}
\name{TL.Gouge}
\alias{TL.Gouge}
\title{TL data of a gouge}
\format{
A matrix 475 x 8 measurement points.
}
\source{
2305181.binx TL grain fin gouge à douille
}
\usage{
data(TLgouge)
}
\description{
An example of actual data from a TL measurement
}
\details{
This data set gives the TL growth curves for gouge
}
\keyword{datasets}
