% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLpack-package.R
\docType{data}
\name{TL.Pan}
\alias{TL.Pan}
\title{TL data of Pan figurine}
\format{
A matrix 475 x 8 measurement points.
}
\source{
L2609193.binx TL grain fin  Pan Br4359 C2RMF78067
}
\usage{
data("TLpan")
}
\description{
An example of actual data from a TL measurement
}
\details{
This data set gives the TL growth curves for Pan figurine Br3459
}
\keyword{datasets}
