% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL.plot.R
\name{TL.plot}
\alias{TL.plot}
\title{TL.plot}
\usage{
TL.plot(File, ech = 1, Doseb0 = 90, Dosea0 = 90, supra = FALSE,
  norm = FALSE, plateau = seq(200, 500), Temp = seq(26, 599),
  NumInv = "", b.check = rep(1, 9), a.check = rep(1, 4),
  sup.check = rep(1, 9))
}
\arguments{
\item{File}{\link{list} (\strong{required}) an output obtained by \link[OSLpack:ReadFile]{OSLpack::ReadFile} function, i.e a list object incluing \linkS4class{Risoe.BINfileData} objects and the names of the corresponding BIN/BINX files}

\item{ech}{\link{numeric} (\strong{with default}) the sample number}

\item{Doseb0}{\link{numeric} (\strong{with default})  the reference of beta irradiation (in seconds)}

\item{Dosea0}{\link{numeric} (\strong{with default})  the reference of alpha irradiation (in seconds)}

\item{supra}{\link{logical} (\strong{with default}) TRUE if supra measurements are inclued. Based on the number of data files, the value is corrected}

\item{norm}{\link{logical} (\strong{with default}) TRUE if the measurements are normalized.}

\item{plateau}{\link{numerical vector} (\strong{with default}) the plateau range for normalization}

\item{Temp}{\link{numerical vector} (\strong{with default}) the temperature range}

\item{NumInv}{\link{string} (\strong{with default}) the inventory number}

\item{b.check}{\link{numerical list} (\strong{with default}) list of beta curves to plot}

\item{a.check}{\link{numerical list} (\strong{with default}) list of alpha curves to plot}

\item{sup.check}{\link{numerical list} (\strong{with default}) list of supralinarity curves to plot}
}
\description{
plot for TL measurements
}
