% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLpack-package.R
\docType{data}
\name{TL.Ser3}
\alias{TL.Ser3}
\title{TL data of an Osiris figurine}
\format{
A matrix 475 x 9 measurement points.
}
\source{
Serapeum E33273 /FGM0873 1510202.binx
}
\usage{
data(TLser3)
}
\description{
An example of actual data from a TL measurement
}
\details{
This data set gives the TL growth curves for Osiris FGM0873
}
\keyword{datasets}
