% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLpack-package.R
\docType{data}
\name{TL.sup.Pan}
\alias{TL.sup.Pan}
\title{Supralinearity data of Pan figurine}
\format{
A matrix 475 x 9 measurement points.
}
\source{
2601201.binx TL supra Tanagra C2RMF77756 2-Base; Pan Br4359 C2RMF78067
}
\usage{
data(TLsupPan)
}
\description{
An example of actual data from a supralinearity measurement
}
\details{
This data set gives the supralinearity for Pan figurine Br3459
}
\keyword{datasets}
