% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLpack-package.R
\docType{data}
\name{TLetru}
\alias{TLetru}
\title{TL data of an etruscan vase}
\format{
A Risoe.BINfileData.object
}
\source{
1601141.bin D171
}
\usage{
data(TLetru)
}
\description{
An example of actual data from a TL measurement
}
\details{
This data set gives the TL growth curves for Etruscan vase D171
}
\keyword{datasets}
