% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLpack-package.R
\docType{data}
\name{TLetru2}
\alias{TLetru2}
\title{TL data of an etruscan vase (2)}
\format{
A Risoe.BINfileData.object
}
\source{
2703141.bin D173
}
\usage{
data(TLetru2)
}
\description{
An example of actual data from a TL measurement
}
\details{
This data set gives the TL growth curves for Etruscan vase D173.

to find how to get Dose, df.T, df.y from Risoe.BINFile file, see example section
}
\examples{
data(TLetru2)
table<-Lum(TLetru2,Doseb=180,alpha=FALSE,supra=FALSE)

B<-table$b
N<-table$n

table.norm<-table$b

for (j in 1:3){
  for (i in 1:3)
  {
    table.norm[,i,j]<-(B[,i,j])/sum(N[seq(355,405),1,(j-1)*3+i])  #norm sur 380-430°C
  }
}

table.données<-cbind(table.norm[,,1],table.norm[,,2],table.norm[,,3])
ii<-c(1,4,7,2,5,8,3,6,9)
table.données<-table.données[seq(1,475),ii]

Dose<-c(0,0,0,180,180,180,360,360,360)
df.T<-matrix(rep(seq(26,500),9),475,9)
df.y<-table.données[,1:9]
}
\keyword{datasets}
