% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lum.R
\name{Lum}
\alias{Lum}
\title{Lum
extracts TL measurements}
\usage{
Lum(
  file,
  ech = 1,
  Doseb0 = 90,
  Dosea0 = 90,
  alpha = TRUE,
  supra = TRUE,
  TypLum = c("TL", "BG"),
  Temp = seq(26, 599)
)
}
\arguments{
\item{file}{\linkS4class{Risoe.BINfileData} (\strong{required}) the BIN/BINX file(s)}

\item{ech}{\link{numeric} (\strong{with default}) the sample number}

\item{Doseb0}{\link{numeric} (\strong{with default})  the reference of beta irradiation (in seconds)}

\item{Dosea0}{\link{numeric} (\strong{with default})  the reference of alpha irradiation (in seconds)}

\item{alpha}{\link{logical} (\strong{with default}) TRUE if alpha measurements are included. Based on the number of data file, the value is corrected}

\item{supra}{\link{logical} (\strong{with default}) TRUE if supra measurements are included. Based on the number of data files, the value is corrected}

\item{TypLum}{\link{character},\link{list} (\strong{with default}) luminescence type "TL" = signal; "BG" = background}

\item{Temp}{\link{numeric},\link{list} (\strong{with default}) the temperature range}
}
\value{
a list object

$alpha

$supra

$b
glowcurve corrected from background for beta irradiation

$a
glowcurve corrected from background for alpha irradiation

$n
glowcurve corrected from background for normalisation

$Bb
uncorrected glowcurve for beta irradiation

$Ba
uncorrected glowcurve for alpha irradiation

$Bn
uncorrected glowcurve for normalisation

$bsup
glowcurve corrected from background for supralinearity irradiation

$nsup
glowcurve corrected from background for supralinearity nomalisation

$Bbsup
uncorrected glowcurve for supralinearity irradiation

$Bnsup
uncorrected glowcurve for supralinearity nomalisation

$CanalTemp
Temperature channel. channel = temperature for Risoe (SystemID=88); channel = Temperature-25 for Lexsyg reader (SYSTEMID=0)
}
\description{
Lum
extracts TL measurements
}
