% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiTL}
\alias{multiTL}
\title{multiTL}
\format{
A list with 4 variables: \code{Dose}, \code{df.T},
\code{df.y}, and \code{iter}
}
\usage{
multiTL
}
\description{
TL example from RLumModel
a dataset containing multiples TL calculated with RLumModel
call function "model_LuminescenceSignals", model = "Bailey2001" test
simulate_sample_history = TRUE
}
\keyword{datasets}
