% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slice5.R
\name{Slice5}
\alias{Slice5}
\title{Slice5}
\usage{
Slice5(
  Dose,
  df.T,
  df.y,
  n.iter,
  n.burnin = n.iter/2,
  n.thin = max(1, floor(n.iter - n.burnin)/10),
  threshold = min(df.y[df.y > 0])
)
}
\arguments{
\item{Dose}{\link{numeric} (\strong{required}) set of the irradiation doses}

\item{df.T}{\link{matrix} (\strong{required}) Luminescence data: the temperatures}

\item{df.y}{\link{matrix} (\strong{required}) Luminescence data; the luminescence signal}

\item{n.iter}{\link{numeric} (\strong{required}) the number of iteration}

\item{n.burnin}{\link{numeric} (\strong{with default})    number of iterations to discard at the beginning (burn in).
Default is n.iter/2, that is, discarding the first half of the simulations.}

\item{n.thin}{\link{numeric} (\strong{with default}) thinning rate. Must be a positive integer. Set n.thin > 1 to save memory and computation time if n.iter is large.
Default is max(1, floor((n.iter-n.burnin) / 10)) which will only thin if there are at least 20 simulations.}

\item{threshold}{\link{numeric} (\strong{with default}) measured point below or equal to the threshold are rejected.
Default is the minimum measured value.}
}
\value{
an (r × 5)-matrix,
\tabular{lll}{
\strong{column} \tab \strong{Type} \tab \strong{Description}\cr
\code{alpha} \tab  \code{numeric} \tab intercept\cr
\code{beta} \tab \code{numeric} \tab slope \cr
\code{sigma2} \tab  \code{numeric} \tab variance\cr
\code{T} \tab \code{numeric} \tab Temperature \cr
\code{De} \tab \code{numeric} \tab 'True' Dose value \cr
}
}
\description{
MC Analysis TL (slice +gibbs) from linear regression
}
\examples{
##load data
if(dev.cur()!=1) dev.off()
data(multiTL, envir = environment())
Dose<-multiTL$Dose
df.T<-multiTL$df.T
df.y<-multiTL$df.y
n.iter<-multiTL$n.iter
test<-Slice5(Dose,df.T,df.y,n.iter=n.iter)
#
if(dev.cur()!=1) dev.off()
data(TLpan, envir = environment())
Dose<-c(0,0,80,80,80,160,160,160)
df.T<-matrix(rep(seq(26,500),8),475,8)
df.y<-TL.Pan[,1:8]
Pan<-Slice5(Dose,df.T,df.y,n.iter=100)
#
\dontrun{
data(TLetru)
table<-Lum(TLetru,Doseb0=360,Dosea=0,alpha=FALSE,supra=FALSE)
B<-table$b
N<-table$n
table.norm<-B
 for (j in 1:3){
   for (i in 1:3)
     {
       table.norm[,i,j]<-(B[,i,j])/sum(N[seq(350,400),1,(j-1)*3+i])
       }
   }
 table.data<-cbind(table.norm[,,1],table.norm[,,2],table.norm[,,3])
 ii<-c(1,4,7,2,5,8,3,6,9)
 table.data<-table.data[seq(1,475),ii]
 Dose<-as.numeric(colnames(table.data))
 df.T<-matrix(rep(seq(26,500),9),475,9)
 df.y<-table.data[,1:9]
 plot(seq(26,500),table.data[,8])
 Slice5(Dose,df.T,df.y,n.iter=10)
 }

}
