% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TL.plot.R
\name{TL.plot}
\alias{TL.plot}
\title{TL.plot}
\usage{
TL.plot(
  file,
  nomFile = "",
  ech = 1,
  Doseb0 = 90,
  Dosea0 = 90,
  supra = FALSE,
  norm = FALSE,
  plateau = seq(200, 500),
  Temp = seq(26, 599),
  NumInv = "",
  b.check = rep(1, 9),
  a.check = rep(1, 4),
  sup.check = rep(1, 9)
)
}
\arguments{
\item{file}{\link{Risoe.BINfileData} (\strong{required}) the BIN/BINX file(s)}

\item{nomFile}{\link{character} (\strong{with default}) name of the BIN/BINX file}

\item{ech}{\link{numeric} (\strong{with default}) the sample number}

\item{Doseb0}{\link{numeric} (\strong{with default})  the reference of beta irradiation (in seconds)}

\item{Dosea0}{\link{numeric} (\strong{with default})  the reference of alpha irradiation (in seconds)}

\item{supra}{\link{logical} (\strong{with default}) TRUE if supra measurements are inclued. Based on the number of data files, the value is corrected}

\item{norm}{\link{logical} (\strong{with default}) TRUE if the measurements are normalized.}

\item{plateau}{\link{numeric},\link{list} (\strong{with default}) the plateau range for normalization}

\item{Temp}{\link{numeric},\link{list} (\strong{with default}) the temperature range}

\item{NumInv}{\link{character} (\strong{with default}) the inventory number}

\item{b.check}{\link{numeric},\link{list} (\strong{with default}) list of beta curves to plot}

\item{a.check}{\link{numeric},\link{list} (\strong{with default}) list of alpha curves to plot}

\item{sup.check}{\link{numeric},\link{list} (\strong{with default}) list of supralinarity curves to plot}
}
\value{
plot figures
}
\description{
plot for TL measurements
}
